Add entity resolver/types/queries
