+++ lib/sanbase/voting/vote.ex
@@ -15,8 +15,6 @@ defmodule Sanbase.Vote do
   alias Sanbase.UserList
   alias Sanbase.Timeline.TimelineEvent
 
-  require Sanbase.Insight.Post.Ecto
-
   @type vote_params :: %{
           :user_id => non_neg_integer(),
           optional(:post_id) => non_neg_integer(),
@@ -135,14 +133,6 @@ defmodule Sanbase.Vote do
     end
   end
 
-  def get_most_voted(entity, opts) do
-    do_get_most_voted(entity, opts)
-  end
-
-  def get_most_recent(entity, opts) do
-    do_get_most_recent(entity, opts)
-  end
-
   def voted_at(entity_type, entity_ids, user_id) when is_integer(user_id) do
     voted_at_query(entity_type, entity_ids, user_id)
     |> Repo.all()
@@ -183,7 +173,7 @@ defmodule Sanbase.Vote do
     # Override nil with -1 so the checks for current user
     # votes will return 0
     user_id = user_id || -1
-    entity_field = deduce_entity_field(entity_type)
+    entity_field = Sanbase.Entity.deduce_entity_field(entity_type)
 
     from(
       vote in entities_query(entity_type, entity_ids),
@@ -203,7 +193,7 @@ defmodule Sanbase.Vote do
   end
 
   defp voted_at_query(entity_type, entity_ids, user_id) do
-    entity_field = deduce_entity_field(entity_type)
+    entity_field = Sanbase.Entity.deduce_entity_field(entity_type)
 
     from(
       vote in entities_query(entity_type, entity_ids),
@@ -216,95 +206,11 @@ defmodule Sanbase.Vote do
   end
 
   defp entities_query(entity_type, entity_ids) do
-    entity_field = deduce_entity_field(entity_type)
+    entity_field = Sanbase.Entity.deduce_entity_field(entity_type)
 
     from(
       vote in __MODULE__,
       where: field(vote, ^entity_field) in ^entity_ids
     )
   end
-
-  defp do_get_most_recent(entity, opts) do
-    {limit, offset} = Sanbase.Utils.Transform.opts_to_limit_offset(opts)
-
-    public_enitiy_ids_query = public_entity_ids_query(entity)
-    entity_module = deduce_entity_module(entity)
-
-    entity_ids =
-      from(
-        entity in entity_module,
-        select: entity.id,
-        where: entity.id in subquery(public_enitiy_ids_query),
-        order_by: [desc: entity.id],
-        limit: ^limit,
-        offset: ^offset
-      )
-      |> Repo.all()
-
-    case entity_module.by_ids(entity_ids, []) do
-      {:ok, result} -> {:ok, Enum.map(result, fn e -> %{entity => e} end)}
-      {:error, error} -> {:error, error}
-    end
-  end
-
-  defp do_get_most_voted(entity, opts) do
-    {limit, offset} = Sanbase.Utils.Transform.opts_to_limit_offset(opts)
-    entity_field = deduce_entity_field(entity)
-
-    # We cannot just find the most voted entity as it could be
-    # made private at some point after getting votes. For this reason
-    # look only at entities that are public. In order to have the same
-    # result for everybody the owner of a private entity does not
-    # get their private entities in the ranking
-    entity_module = deduce_entity_module(entity)
-    public_enitiy_ids_query = public_entity_ids_query(entity)
-
-    entity_ids =
-      from(
-        vote in __MODULE__,
-        right_join: entity in ^entity_module,
-        on: field(vote, ^entity_field) == entity.id,
-        where: entity.id in subquery(public_enitiy_ids_query),
-        group_by: entity.id,
-        select: entity.id,
-        order_by: [desc: coalesce(sum(vote.count), 0), desc: entity.id],
-        limit: ^limit,
-        offset: ^offset
-      )
-
-    entity_ids
-    |> Sanbase.Repo.all()
-    |> entity_module.by_ids([])
-    |> case do
-      {:ok, result} -> {:ok, Enum.map(result, fn e -> %{entity => e} end)}
-      {:error, error} -> {:error, error}
-    end
-  end
-
-  defp public_entity_ids_query(:insight),
-    do: Post.public_entity_ids_query(preload?: false)
-
-  defp public_entity_ids_query(:screener),
-    do: UserList.public_entity_ids_query(is_screener: true)
-
-  defp public_entity_ids_query(:watchlist),
-    do: UserList.public_entity_ids_query(is_screener: false)
-
-  defp public_entity_ids_query(:chart_configuration),
-    do: Chart.Configuration.public_entity_ids_query([])
-
-  defp public_entity_ids_query(:timeline_event),
-    do: TimelineEvent.public_entity_ids_query([])
-
-  defp deduce_entity_module(:insight), do: Post
-  defp deduce_entity_module(:watchlist), do: UserList
-  defp deduce_entity_module(:screener), do: UserList
-  defp deduce_entity_module(:timeline_event), do: TimelineEvent
-  defp deduce_entity_module(:chart_configuration), do: Chart.Configuration
-
-  defp deduce_entity_field(:insight), do: :post_id
-  defp deduce_entity_field(:watchlist), do: :watchlist_id
-  defp deduce_entity_field(:screener), do: :watchlist_id
-  defp deduce_entity_field(:timeline_event), do: :timeline_event_id
-  defp deduce_entity_field(:chart_configuration), do: :chart_configuration_id
 end
