+++ lib/sanbase_web/graphql/resolvers/vote_resolver.ex
@@ -17,14 +17,14 @@ defmodule SanbaseWeb.Graphql.Resolvers.VoteResolver do
     type = Map.get(args, :type)
     page = Map.get(args, :page, 1)
     page_size = Map.get(args, :page_size, 10)
-    Vote.get_most_voted(type, page: page, page_size: page_size)
+    Sanbase.Entity.get_most_voted(type, page: page, page_size: page_size)
   end
 
   def get_most_recent(_root, args, _resolution) do
     type = Map.get(args, :type)
     page = Map.get(args, :page, 1)
     page_size = Map.get(args, :page_size, 10)
-    Vote.get_most_recent(type, page: page, page_size: page_size)
+    Sanbase.Entity.get_most_recent(type, page: page, page_size: page_size)
   end
 
   @doc ~s"""
