+++ lib/sanbase_web/graphql/schema.ex
@@ -42,6 +42,7 @@ defmodule SanbaseWeb.Graphql.Schema do
   import_types(Graphql.ChartConfigurationTypes)
   import_types(Graphql.ClickhouseTypes)
   import_types(Graphql.CommentTypes)
+  import_types(Graphql.EntityTypes)
   import_types(Graphql.EtherbiTypes)
   import_types(Graphql.ExchangeTypes)
   import_types(Graphql.FileTypes)
@@ -87,6 +88,7 @@ defmodule SanbaseWeb.Graphql.Schema do
   import_types(Graphql.Schema.CommentQueries)
   import_types(Graphql.Schema.DiscordQueries)
   import_types(Graphql.Schema.EmailQueries)
+  import_types(Graphql.Schema.EntityQueries)
   import_types(Graphql.Schema.ExchangeQueries)
   import_types(Graphql.Schema.FeaturedQueries)
   import_types(Graphql.Schema.GithubQueries)
@@ -158,6 +160,7 @@ defmodule SanbaseWeb.Graphql.Schema do
     import_fields(:blockchain_metric_queries)
     import_fields(:blockchain_queries)
     import_fields(:comment_queries)
+    import_fields(:entity_queries)
     import_fields(:exchange_queries)
     import_fields(:featured_queries)
     import_fields(:github_queries)
@@ -197,6 +200,7 @@ defmodule SanbaseWeb.Graphql.Schema do
     import_fields(:comment_mutations)
     import_fields(:discord_mutations)
     import_fields(:email_mutations)
+    import_fields(:entity_mutations)
     import_fields(:insight_mutations)
     import_fields(:intercom_mutations)
     import_fields(:project_chart_mutations)
